refactor: split OrdersTable component
